python-service/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
numpy
pillow
torch
opencv-python